Implement end game discovery
We used a macro for our expression due to limitations with
core.match/match; mostly, it does not support unification, just basic
binding of variables.  However, it was a good excuse to use a macro.
Our other disappointment with this solution is that the macro exists
away from the other behaviors related to the board or game.